+++ RELEASE_NOTES.md
@@ -33,6 +33,9 @@ that are used elsewhere in a clause.
 * ADDED: New compiler lint warning for `bagof/3` and `setof/3` where the goal
 argument contains singleton variables.
 
+* ADDED: New compiler lint warning for Logtalk or Prolog top-level shortcuts
+used as directives.
+
 * ADDED: New compiler lint warning for calls to the deprecated `name/2` and
 `current_predicate/2` Prolog built-in predicates.